some viewport have been added
